Data Management Tools (#1)
This is the first complete version of the data acquisition pipeline. It consists of classes which represents a Database and Data Sources, along with additional supporting scripts and modules. The Database class keeps track of the different data sources and the data extracted from them. The DataSource class manages a single data source and is responsible for the ETL (Extract, Transform, Load) process of each source. Currently, all of the data management is fit for a local-machine based database.